requirements.txt
fastapi
pydantic
uvicorn
motor
passlib[bcrypt]
python-jose
python-dotenv
python-multipart